Migrate list-filters-renderer and related
